Show Vaccination modal on row clicks, show Vax card in Encounter
