+++ client/packages/system/src/Vaccination/Components/VaccinationModal.tsx
@@ -34,7 +34,7 @@ import { ArrowRightIcon } from '@mui/x-date-pickers';
 
 interface VaccinationModalProps {
   vaccinationId: string | undefined;
-  encounterId: string;
+  encounterId?: string;
   vaccineCourseDoseId: string;
   isOpen: boolean;
   onClose: () => void;
@@ -128,6 +128,7 @@ export const VaccinationModal = ({
         openBatchModal={openBatchModal}
         draft={draft}
         dose={dose}
+        disabled={!encounterId}
       />
     </Box>
   );
@@ -169,11 +170,13 @@ const VaccinationForm = ({
   dose,
   updateDraft,
   openBatchModal,
+  disabled,
 }: {
   dose?: VaccinationCourseDoseFragment;
   draft: VaccinationDraft;
   updateDraft: (update: Partial<VaccinationDraft>) => void;
   openBatchModal: () => void;
+  disabled: boolean;
 }) => {
   const t = useTranslation('dispensary');
 
@@ -206,6 +209,7 @@ const VaccinationForm = ({
                 });
               }}
               clinicianValue={draft.clinician}
+              disabled={disabled}
             />
           </Grid>
         }
@@ -218,6 +222,7 @@ const VaccinationForm = ({
             value={draft.date}
             onChange={date => updateDraft({ date })}
             sx={{ flex: 1 }}
+            disabled={disabled}
           />
         }
       />
@@ -233,11 +238,13 @@ const VaccinationForm = ({
           value={true}
           control={<Radio />}
           label={t('label.vaccine-given')}
+          disabled={disabled}
         />
         <FormControlLabel
           value={false}
           control={<Radio />}
           label={t('label.vaccine-not-given')}
+          disabled={disabled}
         />
       </RadioGroup>
 
@@ -251,6 +258,7 @@ const VaccinationForm = ({
                 value={draft.itemId ?? ''}
                 onChange={e => updateDraft({ itemId: e.target.value })}
                 sx={{ flex: 1 }}
+                disabled
               />
             }
           />
@@ -293,6 +301,7 @@ const VaccinationForm = ({
                 value={draft.notGivenReason ?? ''}
                 onChange={e => updateDraft({ notGivenReason: e.target.value })}
                 sx={{ flex: 1 }}
+                disabled={disabled}
               />
             }
           />
@@ -311,6 +320,7 @@ const VaccinationForm = ({
               fullWidth
               rows={3}
               style={{ flex: 1 }}
+              disabled={disabled}
             />
           }
         />
